pyproject.toml
[tool.poetry]
name = "django-template"
version = "0.1.0"
description = ""
authors = ["Mehran Kader <kmehran.1106@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "~=3.2"
python-decouple = "^3.4"
logzero = "^1.7.0"
requests = "^2.26.0"
uvicorn = "^0.15.0"
gunicorn = "^20.1.0"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.8.0"
djangorestframework-simplejwt = "^4.8.0"
celery = "^5.1.2"
eventlet = "^0.32.0"
mock = "^4.0.3"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
flake8 = "^3.9.2"
mypy = "^0.910"
isort = "^5.9.3"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38']
experimental_string_processing = true

[tool.isort]
profile = 'black'
multi_line_output = 3
